requirements.txt
# === Frontend Dependencies ===
streamlit
requests
streamlit-extras

# === Backend Dependencies ===
Flask
google-generativeai

